+++ src/containers/NumberGameApp.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import GuessForm from '../components/GuessForm';
-import Moves from '../components/Moves'
-import Game from '../Game';
+import Moves from '../components/NumberMoves'
+import Game from '../NumberGame';
 
 
 class NumberGameApp extends Component {
@@ -28,7 +28,6 @@ class NumberGameApp extends Component {
         if (this.game.isWin() === false) {
             return (
                 <div className="app">
-                    <h1>Game Lobby</h1>
                     <h2>Number Guess Game</h2>
                     <p>Guess number from 0 to 9</p>
                     <GuessForm onSubmit={this.handleGuessSubmit.bind(this)}/>
@@ -39,7 +38,6 @@ class NumberGameApp extends Component {
         } else if (this.game.isWin() === true) {
             return (
                 <div className="app">
-                    <h1>Game Lobby</h1>
                     <h2>Number Guess Game</h2>
                     <p>You won!</p>
                     <h3>Previous moves:</h3>
