+++ test/GameTests.js
@@ -2,7 +2,7 @@
  * Created by minhi_000 on 23.03.2017.
  */
 
-import Game from '../src/Game';
+import Game from '../src/NumberGame';
 
 describe('makeGuess', ()=>{
     it('return `was greater then target` if guessed > than randomed', () =>{
